sanitize lenghts
